+++ ChangeLog.md
@@ -3,6 +3,15 @@ SQL Subcommand change log
 
 ## ?.?.? / ????-??-??
 
+## 0.2.0 / 2017-05-30
+
+* Added ability to use named connection from a per-user configuration
+  file `connections.ini`. This file can be found in:
+  - %LOCALAPPDATA%/Xp-forge.sql/ on Windows
+  - $XDG_CONFIG_HOME/xp-forge.sql/ inside an XDG environment
+  - $HOME/.xp-forge.sql/ otherwise
+  (@thekid)
+
 ## 0.1.0 / 2017-05-30
 
 * Hello World! First release - @thekid